feat: remove items from basket by ajax requests

+++ basketapp/views.py
@@ -45,10 +45,20 @@ def basket_add(request, pk):
 
 @login_required
 def basket_remove(request, pk):
-    basket_record = get_object_or_404(Basket, pk=pk)
-    basket_record.delete()
+    if request.is_ajax():
+        basket_record = get_object_or_404(Basket, pk=pk)
+        basket_record.delete()
 
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        basket_items = Basket.objects.filter(user=request.user).order_by('product__category')
+
+        content = {
+            'basket_items': basket_items,
+        }
+
+        result = render_to_string('basket_list.html', content)
+        return JsonResponse({'result': result})
+    else:
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
 
 
 @login_required
